Refactor + isort
