+++ .gitignore
@@ -18,4 +18,5 @@ tests/example
 tests/example.c
 tests/example.cu
 tests/output.txt
+tests/foo.vbr
 dlmc/
