+++ bench_sparse.py
@@ -1,7 +1,7 @@
-import subprocess
 import os
+import subprocess
 
-from src.codegen import vbr_spmv_codegen, vbr_spmm_codegen
+from src.codegen import vbr_spmm_codegen, vbr_spmv_codegen
 
 BENCHMARK_FREQ = 5
