+++ synthesize_matrices/generate_vbr_matrices_w_all_mostly_dense.sh
