+++ synthesize_matrices/generate_vbr_matrices_w_some_mostly_sparse.sh
